+++ lib/scss_lint/cli.rb
@@ -5,21 +5,36 @@ require 'optparse'
 module SCSSLint
   class CLI
     def initialize(args)
-      opts = OptionParser.new do |opts|
-        opts.banner = 'Usage: scss-lint [scss-files]'
-      end.parse!(args)
+      options = {}
+      parser = OptionParser.new do |opts|
+        opts.banner = "Usage: #{opts.program_name} [options] [scss-files]"
 
-      files = SCSSLint.extract_files_from(opts)
+        opts.separator ''
+        opts.separator 'Common options:'
 
-      runner = Runner.new
-      begin
-        runner.run files
-        report_lints(runner.lints)
-        exit 1 if runner.lints?
-      rescue NoFilesError => ex
-        puts ex.message
-        exit -1
+        opts.on_tail('-h', '--help', 'Show this message') do
+          puts opts.help
+          exit
+        end
+
+        opts.on_tail('-v', '--version', 'Show version') do
+          puts "#{opts.program_name} #{VERSION}"
+          exit
+        end
       end
+
+      begin
+        parser.parse!(args)
+      rescue OptionParser::InvalidOption => ex
+        puts ex, ''
+        puts parser.help
+        exit
+      end
+
+      # Take the rest of the arguments as files/directories
+      options[:files] = args
+
+      run(options)
     end
 
     def report_lints(lints)
@@ -31,6 +46,22 @@ module SCSSLint
         end
 
         puts "#{lint.line} - #{lint.description}"
+      end
+    end
+
+  private
+
+    def run(options)
+      files = SCSSLint.extract_files_from(options[:files])
+
+      runner = Runner.new
+      begin
+        runner.run files
+        report_lints(runner.lints)
+        exit 1 if runner.lints?
+      rescue NoFilesError => ex
+        puts ex.message
+        exit -1
       end
     end
   end