+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix recursive directory scanning
 * Fix specs on Sass gem >= 3.2.6 (shorthand_linter was failing)
 * Add changelog (the thing you're reading)
+* Add command-line flags (e.g. --version, --help)
 
 0.5.2